fixes tests

+++ tests/test_meta_tensor.py
@@ -521,13 +521,13 @@ class TestMetaTensor(unittest.TestCase):
         assert_allclose(obj.as_tensor(), t)
 
     @parameterized.expand(TESTS)
-    def test_array_function(self, device="cpu", _dtype=float):
+    def test_array_function(self, device="cpu", dtype=float):
         a = np.random.RandomState().randn(100, 100)
         b = MetaTensor(a, device=device)
         assert_allclose(np.sum(a), np.sum(b))
         assert_allclose(np.sum(a, axis=1), np.sum(b, axis=1))
         assert_allclose(np.linalg.qr(a), np.linalg.qr(b))
-        c = MetaTensor([1.0, 2.0, 3.0], device=device)
+        c = MetaTensor([1.0, 2.0, 3.0], device=device, dtype=dtype)
         assert_allclose(np.argwhere(c == 1.0).astype(int).tolist(), [[0]])
         assert_allclose(np.concatenate([c, c]), np.asarray([1.0, 2.0, 3.0, 1.0, 2.0, 3.0]))
         if pytorch_after(1, 8, 1):
